% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close.R
\name{close.rscalaBridge}
\alias{close.rscalaBridge}
\title{Close a Scala Bridge}
\usage{
\method{close}{rscalaBridge}(con, ...)
}
\arguments{
\item{con}{A Scala bridge.}

\item{...}{Currently ignored.}
}
\value{
Returns \code{NULL}, invisibly.
}
\description{
Close a Scala Bridge
}
