\name{is.scalaReference}
\alias{is.scalaReference}
\docType{methods}
\title{
Tests whether an object is a Scala reference.
}
\description{
This function returns \code{TRUE} if and only if the object is a Scala reference.
}
\usage{
is.scalaReference(x)
}
\arguments{
    \item{x}{An R object.}
}
\author{
David B. Dahl \email{dahl@stat.byu.edu}
}
\keyword{interface}
\examples{
is.scalaReference(pi)
}

