\name{print}
\alias{print.ScalaInterpreter}
\alias{toString.ScalaInterpreter}
\alias{print.ScalaInterpreterReference}
\alias{toString.ScalaInterpreterReference}
\alias{print.ScalaInterpreterItem}
\alias{toString.ScalaInterpreterItem}
\docType{methods}
\title{
Prints or returns a character representation of an interpreter or a reference.
}
\description{
These functions print or return a character representation of an interpreter or a reference.
}
\usage{
\method{print}{ScalaInterpreter}(x, ...)
\method{toString}{ScalaInterpreter}(x, ...)
\method{print}{ScalaInterpreterReference}(x, ...)
\method{toString}{ScalaInterpreterReference}(x, ...)
\method{print}{ScalaInterpreterItem}(x, ...)
\method{toString}{ScalaInterpreterItem}(x, ...)
}
\arguments{
    \item{x}{An interpreter from the \code{\link{scala}} function or a Scala reference.}
    \item{...}{Currently ignored.}
}
\author{
David B. Dahl \email{dahl@stat.byu.edu}
}
\keyword{interface}

