% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalaAddJAR.R
\name{scalaAddJARs}
\alias{scalaAddJARs}
\title{Add JAR Files to Classpath}
\usage{
scalaAddJARs(bridge, JARs)
}
\arguments{
\item{bridge}{An rscala bridge from the \code{scala} function.}

\item{JARs}{Paths to JAR files, as a character vector.}
}
\value{
Returns \code{NULL}, invisibly.
}
\description{
Add JAR Files to Classpath
}
\examples{
\dontrun{

scalaAddJARs(e, "PATH/TO/jarFileToLoad.jar")
}
}
