% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalaInfo.R
\name{scalaInfo}
\alias{scalaInfo}
\title{Retrieve Information on Scala Installation}
\usage{
scalaInfo(verbose = TRUE)
}
\arguments{
\item{verbose}{Should information on the search be printed?}
}
\value{
Returns a list containing the path to the Scala executable and version information.
}
\description{
This function retrieves information about the installation of Scala.
}
\examples{
scalaInfo()

}
\seealso{
\code{\link{scalaInstall}}
}
