% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalaInstall.R
\name{scalaInstall}
\alias{scalaInstall}
\title{Install Scala}
\usage{
scalaInstall(majorVersion = "2.12")
}
\arguments{
\item{majorVersion}{A valid major Scala release numbers (e.g., "2.12").}
}
\value{
Returns the path to the newly installed executable.
}
\description{
This function installs Scala in the user's \code{~/.rscala} directory.
}
\examples{
\dontrun{

scalaInstall()
}
}
\seealso{
\code{\link{scalaInfo}}
}
