% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalaJARs.R
\name{scalaJARs}
\alias{scalaJARs}
\title{Add JAR Files to Classpath}
\usage{
scalaJARs(JARs, bridge = scalaFindBridge())
}
\arguments{
\item{JARs}{Character vector whose elements are some combination of
individual JAR files or package names which contain embedded JARs.  These
JAR files are added to the runtime classpath.}

\item{bridge}{A Scala bridge from the \code{scala} function.

If the \code{JARs} argument is missing, a character vector of loaded JARs
is returned.}
}
\value{
Returns \code{NULL}, invisibly.
}
\description{
This function is no longer needed but remains until all packages based on
\pkg{rscala} are updated.  It does *not* affect the classpath.
}
\seealso{
\code{\link{scalaFindBridge}}
}
