% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalaMemory.R
\name{scalaMemory}
\alias{scalaMemory}
\title{Get or Set Memory Available to Scala}
\usage{
scalaMemory(x)
}
\arguments{
\item{x}{If the argument is a string (e.g., "8G" or "512M"), the function
sets the default maximum heap size for new instances of Scala bridges
created by the function \code{\link{scala}}.  If the argument is missing,
the current default maximum heap size for new instances is returned.  Set
the argument to \code{NULL} to disable this global option, and therefore
use \pkg{rscala}'s default.  If the argument is a Scala bridge, the
function returns a numeric vector giving the current heap size and the
maximum heap size, in megabytes.}
}
\description{
Depending on the argument type, this function has several uses related to
memory in Scala.
}
\examples{
\dontrun{
scalaMemory("1G")
}

}
\seealso{
\code{\link{scala}}
}
