% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{scalaPackage}
\alias{scalaPackage}
\title{Embed a Scala Bridge in a Package}
\usage{
scalaPackage(packages, assign.callback, mode = "")
}
\arguments{
\item{packages}{A character vector of package names whose JARs should be included in the classpath.}

\item{assign.callback}{A function.}

\item{mode}{A string.  If the package is to have its own bridge, this should
be \code{""}.  If the package is to use the bridge of another package
(e.g., package "foo"), then this should be the fully qualified variable of
that package (e.g., \code{"foo:::s"}).}
}
\value{
Returns \code{NULL}, invisibly.
}
\description{
This function is deprecated.
}
\details{
This function should be called in the \code{\link{.onLoad}} function of a
depending package.

The function makes an rscala bridge available to the functions of the
package. Package developers should call this function in the package's
\code{\link{.onLoad}} function. The \code{\link{scalaPackageUnload}} function
should be called in the package's \code{\link{.onUnload}} function.

This function is deprecated.
}
