% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalaSBTBuildInstallRestart.R
\name{scalaSBTBuildInstallRestart}
\alias{scalaSBTBuildInstallRestart}
\title{Run scalaSBT(), Build, Install, & Restart}
\usage{
scalaSBTBuildInstallRestart()
}
\description{
This function runs \code{scalaSBT()}, builds the source package, installs the package, and restarts R.
}
