% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalaVersionJARs.R
\name{scalaVersionJARs}
\alias{scalaVersionJARs}
\title{JAR Files for Support Scala Versions}
\usage{
scalaVersionJARs()
}
\value{
A list whose names correspond to Scala major versions and whose elements are file system paths.
}
\description{
This function returns a named list whose elements give the file system paths of the JAR files for the supported major versions of Scala.
}
\examples{
scalaVersionJARs()

}
