% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.scalaReference.R
\name{is.scalaReference}
\alias{is.scalaReference}
\title{Test for Scala Reference}
\usage{
is.scalaReference(x)
}
\arguments{
\item{x}{An arbitrary R object.}
}
\value{
Logical indicating whether \code{x} is an rscala reference.
}
\description{
Test for Scala Reference
}
\examples{
is.scalaReference(c(1,2))

}
