\name{rscala-package}
\alias{rscala-package}
\alias{rscala}
\docType{package}
\title{
Bi-directional interface between \R and Scala with callbacks.
}
\description{
This package embeds: i. the \R interpreter in Scala, and ii. the Scala interpreter/compiler in \R (with callbacks to \R).
}
\author{David B. Dahl \email{dahl@stat.byu.edu}}
%\references{}
\seealso{
  \code{\link{scala}}
}
\keyword{package}

