% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalaJARs.R
\name{scalaJARs}
\alias{scalaJARs}
\title{Add JAR Files to Classpath}
\usage{
scalaJARs(JARs, bridge = scalaFindBridge())
}
\arguments{
\item{JARs}{Character vector whose elements are some combination of
individual JAR files or package names which contain embedded JARs.  These
JAR files are added to the runtime classpath.}

\item{bridge}{An rscala bridge from the \code{scala} function.

If the \code{JARs} argument is missing, a character vector of loaded JARs
is returned.}
}
\value{
Returns \code{NULL}, invisibly.
}
\description{
Add JAR Files to Classpath
}
\examples{
\dontrun{

scalaJARs("PATH/TO/jarFileToLoad.jar", e)
}
}
\seealso{
\code{\link{scalaFindBridge}}
}
