% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalaLast.R
\name{scalaLast}
\alias{scalaLast}
\title{Retrieve the Last Scala Computation}
\usage{
scalaLast(bridge)
}
\arguments{
\item{bridge}{An rscala bridge}
}
\description{
This function retrieves the last result from the supplied rscala bridge.
}
\examples{
\donttest{
scala(assign.name='e')      # Implicitly defines the bridge 'e'.
e * "2+3"
scalaLast(e)
close(e)
}

}
