% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalaLast.R
\name{scalaLast}
\alias{scalaLast}
\title{Retrieve the Last Scala Computation}
\usage{
scalaLast(bridge = scalaFindBridge())
}
\arguments{
\item{bridge}{A Scala bridge}
}
\description{
This function retrieves the last result from the supplied Scala bridge.
}
\examples{
\donttest{
s <- scala()
s * "2+3"
scalaLast(s)
close(s)
}

}
\seealso{
\code{\link{scalaFindBridge}}
}
