% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalaPackageUnload.R
\name{scalaPackageUnload}
\alias{scalaPackageUnload}
\title{Clean up an Embedded Scala Bridge of a Package}
\usage{
scalaPackageUnload()
}
\value{
Returns \code{NULL}, invisibly.
}
\description{
This function should be called in the \code{.onUnload} function of a
depending package if, in the corresponding function
\code{\link{scalaPackage}}, the argument \code{mode=""} was used.  It has no
effect otherwise.
}
\examples{
\dontrun{

.onUnload <- function(libname, pkgname) {
  scalaPackageUnload()
}
}

}
\seealso{
\code{\link{scalaPackage}}
}
