% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalaType.R
\name{scalaType}
\alias{scalaType}
\title{Work with Scala Types}
\usage{
scalaType(type)
}
\arguments{
\item{type}{An rscala reference or a character vector of length one giving a Scala type.}
}
\value{
A character vector of length one indicating a Scala type.
}
\description{
Work with Scala Types
}
\examples{
scalaType("Double")
scalaType("D0")
scalaType("Array[String]")
scalaType("S1")

}
